sphinx/requirements.txt
Sphinx
sphinx-rtd-theme
sphinxcontrib-napoleon
pygments
pyyaml
sphinx-argparse
sphinx-copybutton